+++ src/main/java/org/opensearch/knn/index/query/request/MethodParameter.java
@@ -21,7 +21,9 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static org.opensearch.knn.common.KNNConstants.METHOD_PARAMETER_EF_SEARCH;
+import static org.opensearch.knn.common.KNNConstants.METHOD_PARAMETER_NPROBES;
 import static org.opensearch.knn.index.query.KNNQueryBuilder.EF_SEARCH_FIELD;
+import static org.opensearch.knn.index.query.KNNQueryBuilder.NPROBE_FIELD;
 
 /**
  * MethodParameters are engine and algorithm related parameters that clients can pass in knn query
@@ -34,11 +36,7 @@ public enum MethodParameter {
     EF_SEARCH(METHOD_PARAMETER_EF_SEARCH, Version.V_2_16_0, EF_SEARCH_FIELD) {
         @Override
         public Integer parse(Object value) {
-            try {
-                return Integer.parseInt(String.valueOf(value));
-            } catch (final NumberFormatException e) {
-                throw new IllegalArgumentException(METHOD_PARAMETER_EF_SEARCH + " value must be an integer");
-            }
+            return parseInteger(value, METHOD_PARAMETER_EF_SEARCH);
         }
 
         @Override
@@ -47,9 +45,28 @@ public enum MethodParameter {
             if (ef != null && ef > 0) {
                 return null;
             }
-            ;
+
             ValidationException validationException = new ValidationException();
             validationException.addValidationError(METHOD_PARAMETER_EF_SEARCH + " should be greater than 0");
+            return validationException;
+        }
+    },
+
+    NPROBE(METHOD_PARAMETER_NPROBES, Version.V_2_16_0, NPROBE_FIELD) {
+        @Override
+        public Integer parse(Object value) {
+            return parseInteger(value, METHOD_PARAMETER_EF_SEARCH);
+        }
+
+        @Override
+        public ValidationException validate(Object value) {
+            final Integer nprobe = parse(value);
+            if (nprobe != null && nprobe > 0) {
+                return null;
+            }
+
+            ValidationException validationException = new ValidationException();
+            validationException.addValidationError(METHOD_PARAMETER_NPROBES + " should be greater than 0");
             return validationException;
         }
     };
@@ -73,5 +90,13 @@ public enum MethodParameter {
             }
         }
         return PARAMETERS_DIR.get(name);
+    }
+
+    private static Integer parseInteger(Object value, String name) {
+        try {
+            return Integer.parseInt(String.valueOf(value));
+        } catch (final NumberFormatException e) {
+            throw new IllegalArgumentException(name + " value must be an integer");
+        }
     }
 }