+++ src/main/java/org/opensearch/knn/index/util/Faiss.java
@@ -333,7 +333,7 @@ public class Faiss extends NativeLibrary {
     ) {
         super(
             methods,
-            Map.of(METHOD_HNSW, new DefaultHnswContext(), METHOD_IVF, EngineSpecificMethodContext.EMPTY),
+            Map.of(METHOD_HNSW, new DefaultHnswContext(), METHOD_IVF, new DefaultIVFContext()),
             scoreTranslation,
             currentVersion,
             extension